requirements.txt
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.0
decorator==4.4.2
ecdsa==0.18.0
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
mysql-connector-python==8.3.0
numpy==1.26.3
passlib==1.7.4
pillow==10.2.0
proglog==0.1.10
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-ffmpeg-video-streaming==0.1.15
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tqdm==4.66.1
typing_extensions==4.9.0
ua-parser==0.18.0
urllib3==2.1.0
user-agents==2.2.0
uvicorn==0.27.0
watchfiles==0.21.0
websockets==12.0
